Update test fixtures with more realistic data.
Our Geo models and Institution models each have a pk prefixed with the year.
None of our fixtures, however, use the same convention. This updates _several_
fixtures to be more realistic.

This is a good reminder of the challenges of using these sort of database
fixtures in tests. I think it'll be better to switch over to a combination of
model mommy and pytest fixtures.