Rewrite load_hmda command for simplicity.
Previously, the load_hmda command had logic to filter our results based on
state as well as to import a folder of CSVs. This replaces that logic with an
alternative that filters based on which census tracts (and lenders) are in the
database. This is both more concise and runs significantly faster.