Add average number of loans per thousand houses.
We've been displaying the number of LAR for a given record and scaling the
circle size based on the median in a given metro area. Each census tract has a
different number of households, however, so showing the volume can be very
misleading.

This takes a step towards more meaningful results by normalizing the loans
over the number of households. This doesn't change the UI at all, it just adds
the extra data.