Upgrade search library, change URL.
As a step towards allowing us to filter our results by individual counties,
we move the metro search url to be metro-specific. We move its filters into a
separate class which'll be easier to extend. We also move the search form into
a shared (between the intake steps and the map action bar) inclusion tag.
Finally, we upgrade our version of the search library to a version which
allows us to manipulate the search url at query time. Without that, we won't
be able to search for counties by first selecting a state.