Reuse TractFilter in report URLs.
Centralize the logic. This makes reports for counties work without additional
fuss. This _does_ change what's considered an error a bit more (as filters are
a more lax with what input they'll accept). For example, we won't throw a 400
if the user enters a bad year value, we'll just ignore it.