Update webpack version.
Webpack 4 was released in February, along with updates to related modules.
This changeset doesn't make use of all the newest features, but it does update
everything to build our LESS into the expected CSS.

As a by product, it's currently generating some empty JS files corresponding
to the CSS. This is annoying, but should be less of an issue as we build out
the React version of the app.