Add demo version of  SPA app, using React, TS.
As we add more and more state to the slippy map, we've created a fragile,
difficult-to-extend set of scripts. This begins the process of replacing it by
using Redux to centralize state and React to render it. This also introduces
TypeScript to help us describe our functions more precisely (and to write
ES6-like syntax).