Add new sidebar elements in the SPA.
Rather than allowing users to show/hide *every* layer, we group them
into two categories: "choropleth" and "features". Only one of the former
is allowed at a time while "features" are groups of layers which are
shown/hidden together.

This continues to not include tests, as we're still prototyping this.