Use reselect to simplify state management.
Rather than store pre-computed values in the redux store, we can use
reselect to achieve the same result while keeping a much smaller set of
values in the store. This simplifies our Store type significantly, and
centralizes the "update" logic when layers are added or removed.

This also reverts updates to base/map.js, which were accidentally
included in the most recent build.