Add state/actions/reducers for HMDA LAR data.
We need a place in our state tree to track loan data. This adds not only
the slot, but also a bunch of boilerplate around creating the relevant
structures and setting them.

We don't expect every map to display HMDA data, so this is all an
optional configuration. Further, the HMDA configuration may change in
real time, so we don't nest it in the existing, top-level "config"
field.