Display loan circles using a "Marker" component.
For every census tract in which a given lender made loans, we'll display
a circular div with a radius derived from the adjusted volume of loans.
We'll also track a geoid for each loan circle so we have a unique index.

This separates out the LoanCircle div, as we'll probably want to use it
in a key going forward.