Rework state management to three verticals.
Previously, we had kept the majority of our state in top-level keys in
the Store. As we've begun managing more and more, however, that was
getting rather messy. This refactors all of the state to be split into
three verticals: viewport information, mapbox stylesheet/choropleth
information, and LAR data.

In the process, it combines reduces and actions with their associated
state and reworks our LAR data to track multiple lenders, counties, and
metros. This isn't valid on the API side, so the system just grabs the
first of each.