Refactor how filters are stored.
Previously, we stored filters as an array of { id: string, name?:
string }s, using the absence of a name as an indication that it hadn't
been loaded yet. This replaces those elements with a more consistent
structure of the form:
{
  choices: Map<string, string>,
  label: string,
  selected: Set<string>
}
This makes comparisons much simpler and removes all of the indirection
in trying to look at the filter ids.

It's quite a bit of rework, but this should be easier to reason about
going forward.