Add slider to adjust the LAR circle size.
We've tailored the LAR circle size for densely populated areas. When
looking in rural areas, users need to zoom out to see the tracts, which
shrinks the circles. This gives those users a mechanism to scale the
circles back up.

The 1.1 base is somewhat arbitrary; the scaling felt right at this rate,
but we can swap as desired.