Use yattag for template rendering.
As we'll be rendering our reports in HTML (prior to converting them to
PDF), it makes sense to move more template-rendering logic into Python.
While we could use Django's templates to render the reports, it makes
more sense to generate the report markup within the same language we're
using to fetch the data.

This takes a step in that direction by replacing our (one) templated
endpoint with a version built with the yattag library. The only annoying
bit here was that, in the short term, Django Analytical continues to
make sense for generating our analytics tags, so we needed to jump
through a little hoop to continue to render it.